API: move stacking adapters to new subpackage sklearndf.wrapper.stacking
